+++ lib/whois/record/parser/whois.cctld.by.rb
@@ -19,7 +19,7 @@ module Whois
       #   The Example parser for the list of all available methods.
       #
       # @author Aliaksei Kliuchnikau <aliaksei.kliuchnikau@gmail.com>
-      # @since  RELEASE
+      # @since  2.5.0
       class WhoisCctldBy < Base
         include Scanners::Ast