+++ lib/whois/record/scanners/whois.audns.net.au.rb
@@ -16,7 +16,7 @@ module Whois
 
       # Scanner for the whois.audns.net.au record.
       #
-      # @since RELEASE
+      # @since 2.5.0
       class WhoisAudnsNetAu < Base
 
         self.tokenizers += [