+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 
 * CHANGED: Moved scanners from Whois::Record::Parser::Scanners to Whois::Record::Scanners.
 
+* FIXED: .IE availability checking returns incorrect results (GH-155)
+
 
 ## Release 2.4.0